:bug: Fix on utils type conversion int to float

+++ hvsrpy/utils.py
@@ -223,7 +223,7 @@ def parse_hvsrpy_output(fname):
     lookup = {"# Window Length (s)": ("window_length", float),
               "# Total Number of Windows ()": ("total_windows", int),
               "# Frequency Domain Window Rejection Performed ()": ("rejection_bool", bool),
-              "# Number of Standard Deviations Used for Rejection () [n]": ("n_for_rejection", int),
+              "# Number of Standard Deviations Used for Rejection () [n]": ("n_for_rejection", float),
               "# Number of Accepted Windows ()": ("accepted_windows", int),
               "# Distribution of f0 ()": ("distribution_f0", lambda x: x.rstrip()),
               "# Mean f0 (Hz)": ("mean_f0", float),
